requirements.txt
watchdog==3.0.0
PyYAML==6.0.1
python-dotenv==1.0.1
requests==2.31.0
python-telegram-bot==20.7